+++ ci/script.sh
@@ -31,7 +31,7 @@ main() {
 
     case $OPTIONS in
         all)
-            options="--const_generic --strict --derive_more --nightly"
+            options="--const_generic --strict --derive_more --atomics"
             ;;
         *)
             options=$OPTIONS
@@ -46,7 +46,7 @@ main() {
     if [[ "$options" == *"--derive_more"* ]]; then
         echo 'derive_more = "0.99"' >> $td/Cargo.toml
     fi
-    if [[ "$options" == *"--nightly"* ]]; then
+    if [[ "$options" == *"--atomics"* ]]; then
         echo 'portable-atomic = { version = "0.3.15", default-features = false }' >> $td/Cargo.toml
     fi
     echo '[profile.dev]' >> $td/Cargo.toml